+++ python/ray/tests/BUILD
@@ -90,6 +90,7 @@ py_test_module_list(
     "test_multinode_failures_2.py",
     "test_multiprocessing.py",
     "test_object_assign_owner.py",
+    "test_placement_group.py",
     "test_placement_group_2.py",
     "test_placement_group_3.py",
     "test_ray_init.py",
@@ -180,7 +181,6 @@ py_test_module_list(
     "test_failure_3.py",
     "test_chaos.py",
     "test_reference_counting_2.py",
-    "test_placement_group.py",
   ],
   size = "large",
   extra_srcs = SRCS,
